pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "schema_gen_ai"
version = "0.1.0"
description = "A Python tool to generate Pydantic response schemas using various LLM APIs"
readme = "README.md"
authors = [{name = "Gokulakrishnan Anand", email = "gokula04@gmail.com"}]
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
        'pydantic>=2.11.7',
        'openai>=2.6.1',
        'anthropic>=0.71.0',
        'google-genai>=1.46.0',
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
Homepage = "https://github.com/Gokul-A/schema_gen_ai"
Repository = "https://github.com/Gokul-A/schema_gen_ai"